Add support for commiting to Git repositories automatically.
This commit adds a number of new command-line options. Most notable of
these is the ability to pass `--scm`, which enables the source code
management feature. This feature automatically looks for a SCM checkout
at the location of the `--download-directory` and commits the newly
downloaded content as a revision atop the existing repository. If a
repository does not exist, it will be automatically intialized.

The only SCM backend currently supported is Git, so passing `--scm` and
`--scm git` are equivalent.

The `--scm-args` option can be given a single (quoted) argument that
will be passed to the SCM backend chosen. This allows for (almost)
arbitrarily complex commit options to be provided and should make
integration with other SCM systems easier in the future, too. Beware
that parsing multiple options from this single string is currently
somewhat naive.

This commit further adds support for arbitrary `wget` options using a
similar `--wget-args` option.

Documentation has been updated, as well, with plenty more code comments
and a number of additional examples in the README and `usage()` output.